Fix/iso always 0 (#60)
* docs: chore

* fix: getExifField returns int64.

+++ src/viewer/cmd/tagger/tagger.go
@@ -81,7 +81,7 @@ func FillMetadataByExif(e *exif.Exif) (*model.Metadata, error) {
 		fl = int(_fl)
 	}
 	iso := 0
-	if _iso, ok := getExifField(e, exif.ISOSpeedRatings).(int); ok {
+	if _iso, ok := getExifField(e, exif.ISOSpeedRatings).(int64); ok {
 		iso = int(_iso)
 	}
